removed unused bloat
