+++ Quaerere_client/.expo/packager-info.json
@@ -1,8 +1,8 @@
 {
   "devToolsPort": 19002,
   "expoServerPort": null,
-  "packagerPort": null,
-  "webpackServerPort": null,
+  "packagerPort": 19000,
+  "webpackServerPort": 19006,
   "expoServerNgrokUrl": null,
   "packagerNgrokUrl": null,
   "ngrokPid": null,
