--- Quaerere_client/assets/css/base.css
@@ -1,16 +1,0 @@
- /* Style the navigation menu */
- #root{
-   margin: auto;
-   width: 50%;
- }
- .topnav {
-    overflow: hidden;
-    background-color: #333;
-    position: relative;
-  }
-  
-  /* Hide the links inside the navigation menu (except for logo/home) */
-  .topnav #myLinks {
-    display: none;
-  }
-
